requirements.txt
Django==1.8.6
djangorestframework==3.3.1
djangorestframework-jwt==1.7.2
PyJWT==1.4.0
wheel==0.24.0
